# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EasyStay Documentation",
        "description": "Ceci est la documentation de EasyStay.",
        "contact": {
            "email": "haroldavademe0@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservation/storeReserv": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Créer une réservation pour une chambre",
                "description": "Cette API permet de créer une réservation pour une chambre en fonction des dates sélectionnées.",
                "operationId": "storeReservation",
                "parameters": [
                    {
                        "name": "chambre_id",
                        "in": "query",
                        "description": "Identifiant de la chambre à réserver",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Identifiant de l'utilisateur qui fait la réservation (facultatif si un email est fourni)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email de l'utilisateur (facultatif si un user_id est fourni)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "user@example.com"
                        }
                    },
                    {
                        "name": "date_deb",
                        "in": "query",
                        "description": "Date de début de la réservation (format : YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-25"
                        }
                    },
                    {
                        "name": "date_fin",
                        "in": "query",
                        "description": "Date de fin de la réservation (format : YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-30"
                        }
                    },
                    {
                        "name": "nmb_per",
                        "in": "query",
                        "description": "Nombre de personnes pour la réservation",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "chambre_id": {
                                        "description": "Identifiant de la chambre à réserver",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "description": "Identifiant de l'utilisateur qui fait la réservation (facultatif si un email est fourni)",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "email": {
                                        "description": "Email de l'utilisateur (facultatif si un user_id est fourni)",
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "date_deb": {
                                        "description": "Date de début de la réservation (format : YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-11-25"
                                    },
                                    "date_fin": {
                                        "description": "Date de fin de la réservation (format : YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-11-30"
                                    },
                                    "nmb_per": {
                                        "description": "Nombre de personnes pour la réservation",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réservation créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Réservation réussie."
                                        },
                                        "reservation": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                },
                                                "chambre_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "date_deb": {
                                                    "type": "string",
                                                    "example": "2024-11-25"
                                                },
                                                "date_fin": {
                                                    "type": "string",
                                                    "example": "2024-11-30"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "CONFIRMED"
                                                },
                                                "nmb_per": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "tarif_app": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 500
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation ou disponibilité de la chambre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La chambre n'est pas disponible pour ces dates."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chambre non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Chambre non trouvée."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé (par exemple, pour un utilisateur non authentifié)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès refusé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}